+++ recipes/default.rb
@@ -18,11 +18,11 @@
 # limitations under the License.
 #
 
-include_recipe 'nginx::ohai_plugin'
+include_recipe 'ljandrew_nginx::ohai_plugin'
 
 case node['nginx']['install_method']
 when 'source'
-  include_recipe 'nginx::source'
+  include_recipe 'ljandrew_nginx::source'
 when 'package'
   case node['platform']
   when 'redhat','centos','scientific','amazon','oracle'
@@ -33,7 +33,7 @@ when 'package'
     supports :status => true, :restart => true, :reload => true
     action :enable
   end
-  include_recipe 'nginx::commons'
+  include_recipe 'ljandrew_nginx::commons'
 end
 
 service 'nginx' do