+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## v1.0.2:
+
+* [COOK-1636] - relax the version constraint on ohai
+
+## v1.0.0:
+
+* [COOK-913] - defaults for gzip cause warning on service restart
+* [COOK-1020] - duplicate MIME type
+* [COOK-1269] - add passenger module support through new recipe
+* [COOK-1306] - increment nginx version to 1.2 (now 1.2.3)
+* [COOK-1316] - default site should not always be enabled
+* [COOK-1417] - resolve errors preventing build from source
+* [COOK-1483] - source prefix attribute has no effect
+* [COOK-1484] - source relies on /etc/sysconfig
+* [COOK-1511] - add support for naxsi module
+* [COOK-1525] - nginx source is downloaded every time
+* [COOK-1526] - nginx_site does not remove sites
+* [COOK-1527] - add `http_echo_module` recipe
+
 ## v0.101.6:
 
 Erroneous cookbook upload due to timeout.